requirements.txt
astroid==3.3.11
blinker==1.9.0
click==8.1.8
dill==0.4.0
Flask==3.1.2
flask-cors==6.0.1
importlib_metadata==8.7.0
isort==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
platformdirs==4.4.0
pylint==3.3.9
tomli==2.3.0
tomlkit==0.13.3
typing_extensions==4.15.0
Werkzeug==3.1.4
zipp==3.23.0
python-dotenv==1.0.0
google-genai
celery==5.4.0
redis==5.0.1
